Implament FF in IO blocks.
There are three registers in IO blocks: IREG, OREG, and TREG. Not in all
of them, but only where there is IOLOGIC because these registers are
part of it.

Hence the first limitation: if FF is placed in an IO block, the rest of
the IOLOGIC cannot be used.

The second limitation is related to the fact that all three registers
have common wires CLK, ClockEnable and LocalSetReset and this dictates
the types of flipflop's - let's say DFF (in Gowin's implementation)
requires a constant connection CE to VCC, so you can not use in the same
IO block DFFE, which CE is a signal.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>